reporting_app/requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Django==3.1.4
django-celery-beat==2.1.0
django-environ==0.4.5
django-timezone-field==4.1.1
djangorestframework==3.12.2
idna==2.10
kombu==5.0.2
prompt-toolkit==3.0.8
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
redis==3.5.3
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
